usigned long -> uint32_t

+++ CRC32.cpp
@@ -10,13 +10,13 @@ CRC32::~CRC32()
 }
 
 void CRC32::make_crc_table() {
-  unsigned long POLYNOMIAL = 0xEDB88320;
-  unsigned long remainder;
+  uint32_t POLYNOMIAL = 0xEDB88320;
+  uint32_t remainder;
   unsigned char b = 0;
   do {
     // Start with the data byte
     remainder = b;
-    for (unsigned long bit = 8; bit > 0; --bit)
+    for (uint32_t bit = 8; bit > 0; --bit)
     {
       if (remainder & 1)
         remainder = (remainder >> 1) ^ POLYNOMIAL;
@@ -27,8 +27,8 @@ void CRC32::make_crc_table() {
     } while(0 != ++b);
 }
 
-unsigned long CRC32::crc(unsigned char *p, size_t n) {
-  unsigned long crc = 0xfffffffful;
+uint32_t CRC32::crc(unsigned char *p, size_t n) {
+  uint32_t crc = 0xfffffffful;
   size_t i;
   for(i = 0; i < n; i++)
     crc = crcTable[*p++ ^ (crc&0xff)] ^ (crc>>8);
